backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyPDF2==3.0.1
openai==1.58.1
google-generativeai==0.8.3
requests==2.32.3
gunicorn==23.0.0
whitenoise==6.8.2
